#milestone
ok as of now it setup that it there is no more bunny cdn and it wont be used. front end code is directly hosted in webflow, but ill leave the account open so in the future if i need it i can quickly configure it and quickly configure it. previous commit i think 140 still this code for the github action to automatically deploy.

i did the above so now the api keys are hidden for supabase. increases security as i also have rls configured

now i need to setup email and make sure all the is configured for both ends, and make some slight adjusts to how the data is formatted and we should be good from here